requirements.txt
django==2.2.7
djangorestframework==3.9.3
pytest-django==3.5.1
psycopg2==2.8.4